% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_foreign_keys.R
\name{create_foreign_keys}
\alias{create_foreign_keys}
\title{Create foreign keys for the tables in the PostgreSQL database}
\usage{
create_foreign_keys(conn, schema_name, relationships, uploaded_tables)
}
\arguments{
\item{conn}{A DBI connection object to the PostgreSQL database.}

\item{schema_name}{A character string specifying the name of the schema.}

\item{relationships}{A list of lists specifying the relationships between the tables.}
}
\description{
Create foreign keys for the tables in the PostgreSQL database
}
