% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{Create a new schema in the PostgreSQL database}
\usage{
create_schema(conn, schema_name)
}
\arguments{
\item{conn}{A DBI connection object to the PostgreSQL database.}

\item{schema_name}{A character string specifying the name of the schema to be created.}
}
\description{
Create a new schema in the PostgreSQL database
}
