% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_id_variables.R
\name{identify_id_variables}
\alias{identify_id_variables}
\title{Identify ID variables for each table type}
\usage{
identify_id_variables(table_name)
}
\arguments{
\item{table_name}{A character string representing the full table name.}
}
\value{
A list of character strings representing the ID variables for the table type.
}
\description{
Identify ID variables for each table type
}
