% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_relationships.R
\name{identify_relationships}
\alias{identify_relationships}
\title{Identify relationships between the dataframes in dhs_data_list}
\usage{
identify_relationships(dhs_data_list)
}
\arguments{
\item{dhs_data_list}{A named list of dataframes containing the DHS data.}
}
\value{
A list of lists specifying the relationships between the tables.
}
\description{
Identify relationships between the dataframes in dhs_data_list
}
