% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataframe.R
\name{split_dataframe}
\alias{split_dataframe}
\title{Split a dataframe with more than 1600 columns into multiple dataframes}
\usage{
split_dataframe(df, id_vars)
}
\arguments{
\item{df}{A dataframe to be split.}

\item{id_vars}{A character vector specifying the ID variables to be included in each split dataframe.}
}
\value{
A list of dataframes.
}
\description{
Split a dataframe with more than 1600 columns into multiple dataframes
}
