% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_exists.R
\name{table_exists}
\alias{table_exists}
\title{Check if a table exists in the PostgreSQL database}
\usage{
table_exists(conn, schema_name, table_name)
}
\arguments{
\item{conn}{A connection object to the PostgreSQL database.}

\item{schema_name}{A character string representing the schema name.}

\item{table_name}{A character string representing the table name.}
}
\value{
A logical value indicating whether the table exists in the PostgreSQL database.
}
\description{
Check if a table exists in the PostgreSQL database
}
