% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_dataframe_to_postgres.R
\name{upload_dataframe_to_postgres}
\alias{upload_dataframe_to_postgres}
\title{Upload a dataframe to the PostgreSQL database}
\usage{
upload_dataframe_to_postgres(conn, schema_name, table_name, df)
}
\arguments{
\item{conn}{A DBI connection object to the PostgreSQL database.}

\item{schema_name}{A character string specifying the name of the schema to be created.}

\item{table_name}{A character string specifying the name of the table to be created.}

\item{df}{A dataframe to be uploaded to the PostgreSQL database.}
}
\description{
Upload a dataframe to the PostgreSQL database
}
